update QL menu, NV, TK, DH.

+++ app/src/main/java/com/example/dacn/View/KhachHangActivity.java
@@ -147,7 +147,7 @@ public class KhachHangActivity extends AppCompatActivity implements ProductAdapt
                     product.getMaSanPham(), // Mã sản phẩm
                     product.getTenSanPham(), // Tên sản phẩm
                     product.getHinhAnh(), // URL hình ảnh
-                    product.getGia(), // Giá sản phẩm
+                    (double) product.getGia(), // Giá sản phẩm
 
                     1 // Số lượng mặc định là 1
             );
